// app/schemas/auth.ts
/** @format */

import z from "zod";

export const signUpSchema = z.object({
  name: z.string().min(3, "Name must be at least 3 characters long").max(50),

  email: z.email(),
  password: z.string().min(8).max(30),
});

export const loginSchema = z.object({
  email: z.email(),
  password: z.string().min(8).max(30),
});